move chunker to borg.algorithms
